+++ plugin/main/src/kotlinx/benchmark/gradle/BenchmarksPlugin.kt
@@ -120,6 +120,7 @@ constructor(
                 is NativeBenchmarkTarget -> processNativeCompilation(config)
             }
         }
+        extension.checkConflictingJmhVersions()
     }
 
     private fun configureBenchmarkTaskConventions(